refactor(swingset): shared `extractPresenceIfPresent`
