+++ packages/SwingSet/src/kernel/kernel.js
@@ -11,7 +11,7 @@ import makeKernelKeeper from './state/kernelKeeper.js';
 import { kdebug, kdebugEnable, legibilizeMessageArgs } from '../lib/kdebug.js';
 import { insistKernelType, parseKernelSlot } from './parseKernelSlots.js';
 import { parseVatSlot } from '../lib/parseVatSlots.js';
-import { insistCapData } from '../lib/capdata.js';
+import { extractPresenceIfPresent, insistCapData } from '../lib/capdata.js';
 import { insistMessage, insistVatDeliveryResult } from '../lib/message.js';
 import { insistDeviceID, insistVatID } from '../lib/id.js';
 import { makeKernelQueueHandler } from './kernelQueue.js';
@@ -406,25 +406,6 @@ export default function buildKernel(
     return deliverAndLogToVat(vatID, kd, vd);
   }
 
-  function extractPresenceIfPresent(data) {
-    const body = JSON.parse(data.body);
-    if (
-      body &&
-      typeof body === 'object' &&
-      body['@qclass'] === 'slot' &&
-      body.index === 0
-    ) {
-      if (data.slots.length === 1) {
-        const slot = data.slots[0];
-        const { type } = parseKernelSlot(slot);
-        if (type === 'object') {
-          return slot;
-        }
-      }
-    }
-    return null;
-  }
-
   /**
    *
    * @param { RunQueueEventNotify } message
@@ -787,7 +768,7 @@ export default function buildKernel(
     const kp = kernelKeeper.getKernelPromise(target);
     switch (kp.state) {
       case 'fulfilled': {
-        const presence = extractPresenceIfPresent(kp.data);
+        const presence = extractPresenceIfPresent(kp.data, parseKernelSlot);
         if (presence) {
           return send(presence);
         }
